#25: refactored and updated labels
