+++ web/modules/custom/geolayer/modules/map/config/install/geolayer_map.map_behavior.geojson.yml
@@ -1,0 +1,11 @@
+langcode: en
+status: true
+dependencies:
+  enforced:
+    module:
+      - geolayer_map
+id: geojson
+label: Geojson
+description: 'Renders geojson using geolayer-map.'
+behaviors: {  }
+options: {  }
